minor correction

+++ LSDChiNetwork.hpp
@@ -827,7 +827,7 @@ class LSDChiNetwork
     /// @return vector of vectors with m means
     /// @ author DTM
     /// @date 24/03/15
-    vector< vector<float> > get_node_indices()
+    vector< vector<int> > get_node_indices()
       { return node_indices; }
     /// @brief This gets the chi coordinates for the channel network
     /// @return vector of vectors with m means
